requirements.txt
Flask==2.3.3
Flask-SQLAlchemy==3.0.5
Flask-Login==0.6.2
Flask-WTF==1.1.1
WTForms==3.0.1
Werkzeug==2.3.7
python-dotenv==1.0.0
pytz==2023.3
swisseph==2.10.03
requests==2.31.0
openai==0.27.8
bcrypt==4.0.1
email-validator==2.0.0
matplotlib==3.7.2
Pillow==10.0.0
numpy==1.24.3
pandas==2.0.3
python-dateutil==2.8.2
gunicorn==21.2.0
# Additional packages for robust astronomical calculations
geopy==2.4.0
timezonefinder==6.2.0
astral==3.2